fix: isolate exclusive linux functions and add darwin implementation
